updates graph api to use tck entities
