+++ artemis/artemis-graph/src/test/java/org/eclipse/jnosql/artemis/graph/AbstractGraphTemplateTest.java
@@ -18,15 +18,15 @@ import jakarta.nosql.NonUniqueResultException;
 import jakarta.nosql.mapping.EntityNotFoundException;
 import jakarta.nosql.mapping.IdNotFoundException;
 import jakarta.nosql.mapping.PreparedStatement;
+import jakarta.nosql.tck.entities.Animal;
+import jakarta.nosql.tck.entities.Book;
+import jakarta.nosql.tck.entities.Person;
+import jakarta.nosql.tck.entities.WrongEntity;
 import org.apache.tinkerpop.gremlin.structure.Direction;
 import org.apache.tinkerpop.gremlin.structure.Edge;
 import org.apache.tinkerpop.gremlin.structure.Graph;
 import org.apache.tinkerpop.gremlin.structure.Transaction;
 import org.apache.tinkerpop.gremlin.structure.Vertex;
-import org.eclipse.jnosql.artemis.graph.model.Animal;
-import org.eclipse.jnosql.artemis.graph.model.Book;
-import org.eclipse.jnosql.artemis.graph.model.Person;
-import org.eclipse.jnosql.artemis.graph.model.WrongEntity;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.AfterEach;
