+++ artemis/artemis-graph/src/test/java/org/eclipse/jnosql/artemis/graph/BookTemplateTest.java
@@ -15,14 +15,12 @@
 package org.eclipse.jnosql.artemis.graph;
 
 
+import jakarta.nosql.tck.entities.Book;
+import jakarta.nosql.tck.test.CDIExtension;
 import org.apache.tinkerpop.gremlin.structure.Graph;
 import org.apache.tinkerpop.gremlin.structure.Transaction;
 import org.apache.tinkerpop.gremlin.structure.Transaction.Status;
-import org.eclipse.jnosql.artemis.graph.model.Book;
-import org.eclipse.jnosql.artemis.graph.model.BookTemplate;
-import jakarta.nosql.tck.test.CDIExtension;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 
 import javax.inject.Inject;
 import java.util.concurrent.atomic.AtomicReference;
