+++ artemis/artemis-graph/src/test/java/org/eclipse/jnosql/artemis/graph/DefaultValueMapTraversalTest.java
@@ -15,7 +15,7 @@
 package org.eclipse.jnosql.artemis.graph;
 
 import jakarta.nosql.NonUniqueResultException;
-import org.eclipse.jnosql.artemis.graph.model.Person;
+import jakarta.nosql.tck.entities.Person;
 import jakarta.nosql.tck.test.CDIExtension;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
