+++ artemis/artemis-graph/src/test/java/org/eclipse/jnosql/artemis/graph/GremlinExecutorTest.java
@@ -14,12 +14,12 @@
  */
 package org.eclipse.jnosql.artemis.graph;
 
+import jakarta.nosql.tck.entities.Person;
+import jakarta.nosql.tck.test.CDIExtension;
 import org.apache.tinkerpop.gremlin.structure.Edge;
 import org.apache.tinkerpop.gremlin.structure.Graph;
 import org.apache.tinkerpop.gremlin.structure.T;
 import org.apache.tinkerpop.gremlin.structure.Vertex;
-import org.eclipse.jnosql.artemis.graph.model.Person;
-import jakarta.nosql.tck.test.CDIExtension;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
