--- artemis/artemis-graph/src/test/java/org/eclipse/jnosql/artemis/graph/model/Animal.java
@@ -1,71 +1,0 @@
-/*
- *  Copyright (c) 2017 Otávio Santana and others
- *   All rights reserved. This program and the accompanying materials
- *   are made available under the terms of the Eclipse Public License v1.0
- *   and Apache License v2.0 which accompanies this distribution.
- *   The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v10.html
- *   and the Apache License v2.0 is available at http://www.opensource.org/licenses/apache2.0.php.
- *
- *   You may elect to redistribute this code under either of these licenses.
- *
- *   Contributors:
- *
- *   Otavio Santana
- */
-package org.eclipse.jnosql.artemis.graph.model;
-
-import jakarta.nosql.mapping.Column;
-import jakarta.nosql.mapping.Entity;
-import jakarta.nosql.mapping.Id;
-
-import java.util.Objects;
-
-@Entity
-public class Animal {
-
-
-    @Id
-    private Long id;
-
-    @Column
-    private String name;
-
-    Animal() {
-    }
-
-    public Animal(String name) {
-        this.name = name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (!(o instanceof Animal)) {
-            return false;
-        }
-        Animal animal = (Animal) o;
-        return Objects.equals(id, animal.id) &&
-                Objects.equals(name, animal.name);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, name);
-    }
-
-    @Override
-    public String toString() {
-        return  "Animal{" + "name='" + name + '\'' +
-                '}';
-    }
-}
