--- artemis/artemis-graph/src/test/java/org/eclipse/jnosql/artemis/graph/model/Book.java
@@ -1,114 +1,0 @@
-/*
- *  Copyright (c) 2017 Otávio Santana and others
- *   All rights reserved. This program and the accompanying materials
- *   are made available under the terms of the Eclipse Public License v1.0
- *   and Apache License v2.0 which accompanies this distribution.
- *   The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v10.html
- *   and the Apache License v2.0 is available at http://www.opensource.org/licenses/apache2.0.php.
- *
- *   You may elect to redistribute this code under either of these licenses.
- *
- *   Contributors:
- *
- *   Otavio Santana
- */
-package org.eclipse.jnosql.artemis.graph.model;
-
-import jakarta.nosql.mapping.Column;
-import jakarta.nosql.mapping.Entity;
-import jakarta.nosql.mapping.Id;
-
-import java.util.Objects;
-
-@Entity
-public class Book {
-
-    @Id
-    private Long id;
-
-    @Column
-    private String name;
-
-    @Column
-    private Integer age;
-
-
-    Book() {
-    }
-
-    Book(Long id, String name, Integer age) {
-        this.id = id;
-        this.name = name;
-        this.age = age;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public Integer getAge() {
-        return age;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (!(o instanceof Book)) {
-            return false;
-        }
-        Book book = (Book) o;
-        return Objects.equals(id, book.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(id);
-    }
-
-    @Override
-    public String toString() {
-        return "Book{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", age=" + age +
-                '}';
-    }
-
-    public static BookBuilder builder() {
-        return new BookBuilder();
-    }
-
-    public static class BookBuilder {
-        private String name;
-        private Integer age;
-        private Long id;
-
-        private BookBuilder() {
-        }
-
-        public BookBuilder withName(String name) {
-            this.name = name;
-            return this;
-        }
-
-        public BookBuilder withAge(Integer age) {
-            this.age = age;
-            return this;
-        }
-
-        public BookBuilder withId(Long id) {
-            this.id = id;
-            return this;
-        }
-
-        public Book build() {
-            return new Book(id, name, age);
-        }
-    }
-}
