--- artemis/artemis-graph/src/test/java/org/eclipse/jnosql/artemis/graph/model/Job.java
@@ -1,64 +1,0 @@
-/*
- *  Copyright (c) 2017 Otávio Santana and others
- *   All rights reserved. This program and the accompanying materials
- *   are made available under the terms of the Eclipse Public License v1.0
- *   and Apache License v2.0 which accompanies this distribution.
- *   The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v10.html
- *   and the Apache License v2.0 is available at http://www.opensource.org/licenses/apache2.0.php.
- *
- *   You may elect to redistribute this code under either of these licenses.
- *
- *   Contributors:
- *
- *   Otavio Santana
- */
-package org.eclipse.jnosql.artemis.graph.model;
-
-
-import jakarta.nosql.mapping.Column;
-import jakarta.nosql.mapping.Embeddable;
-
-import java.util.Objects;
-
-@Embeddable
-public class Job {
-
-    @Column
-    private String description;
-
-    @Column
-    private String city;
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Job job = (Job) o;
-        return Objects.equals(description, job.description);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(description);
-    }
-}
