--- artemis/artemis-graph/src/test/java/org/eclipse/jnosql/artemis/graph/model/Money.java
@@ -1,70 +1,0 @@
-/*
- *  Copyright (c) 2017 Otávio Santana and others
- *   All rights reserved. This program and the accompanying materials
- *   are made available under the terms of the Eclipse Public License v1.0
- *   and Apache License v2.0 which accompanies this distribution.
- *   The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v10.html
- *   and the Apache License v2.0 is available at http://www.opensource.org/licenses/apache2.0.php.
- *
- *   You may elect to redistribute this code under either of these licenses.
- *
- *   Contributors:
- *
- *   Otavio Santana
- */
-package org.eclipse.jnosql.artemis.graph.model;
-
-
-import java.math.BigDecimal;
-import java.util.Objects;
-
-public class Money {
-
-    private final String currency;
-
-    private final BigDecimal value;
-
-    public Money(String currency, BigDecimal value) {
-        this.currency = currency;
-        this.value = value;
-    }
-
-    public String getCurrency() {
-        return currency;
-    }
-
-    public BigDecimal getValue() {
-        return value;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Money money = (Money) o;
-        return Objects.equals(currency, money.currency) &&
-                Objects.equals(value.doubleValue(), money.value.doubleValue());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(currency, value.doubleValue());
-    }
-
-    @Override
-    public String toString() {
-        return currency + " " + value.toString();
-    }
-
-    public static Money parse(String dbData) {
-        String[] values = dbData.split(" ");
-        String currency = values[0];
-        BigDecimal value = BigDecimal.valueOf(Double.valueOf(values[1]));
-        return new Money(currency, value);
-    }
-}
-
