--- artemis/artemis-graph/src/test/java/org/eclipse/jnosql/artemis/graph/model/MoneyConverter.java
@@ -1,32 +1,0 @@
-/*
- *  Copyright (c) 2017 Otávio Santana and others
- *   All rights reserved. This program and the accompanying materials
- *   are made available under the terms of the Eclipse Public License v1.0
- *   and Apache License v2.0 which accompanies this distribution.
- *   The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v10.html
- *   and the Apache License v2.0 is available at http://www.opensource.org/licenses/apache2.0.php.
- *
- *   You may elect to redistribute this code under either of these licenses.
- *
- *   Contributors:
- *
- *   Otavio Santana
- */
-package org.eclipse.jnosql.artemis.graph.model;
-
-
-import jakarta.nosql.mapping.AttributeConverter;
-
-public class MoneyConverter implements AttributeConverter<Money, String> {
-
-
-    @Override
-    public String convertToDatabaseColumn(Money attribute) {
-        return attribute.toString();
-    }
-
-    @Override
-    public Money convertToEntityAttribute(String dbData) {
-        return Money.parse(dbData);
-    }
-}
