--- artemis/artemis-graph/src/test/java/org/eclipse/jnosql/artemis/graph/model/Person.java
@@ -1,110 +1,0 @@
-/*
- *  Copyright (c) 2017 Otávio Santana and others
- *   All rights reserved. This program and the accompanying materials
- *   are made available under the terms of the Eclipse Public License v1.0
- *   and Apache License v2.0 which accompanies this distribution.
- *   The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v10.html
- *   and the Apache License v2.0 is available at http://www.opensource.org/licenses/apache2.0.php.
- *
- *   You may elect to redistribute this code under either of these licenses.
- *
- *   Contributors:
- *
- *   Otavio Santana
- */
-package org.eclipse.jnosql.artemis.graph.model;
-
-
-import jakarta.nosql.mapping.Column;
-import jakarta.nosql.mapping.Entity;
-import jakarta.nosql.mapping.Id;
-import jakarta.nosql.mapping.MappedSuperclass;
-
-import java.util.List;
-import java.util.Objects;
-
-@Entity
-@MappedSuperclass
-public class Person {
-
-    @Id
-    private Long id;
-
-    @Column
-    private String name;
-
-    @Column
-    private int age;
-
-    @Column
-    private List<String> phones;
-
-    private String ignore;
-
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public List<String> getPhones() {
-        return phones;
-    }
-
-    public String getIgnore() {
-        return ignore;
-    }
-
-    public boolean isAdult() {
-        return age > 21;
-    }
-
-    Person() {
-    }
-
-    Person(Long id, String name, int age, List<String> phones, String ignore) {
-        this.id = id;
-        this.name = name;
-        this.age = age;
-        this.phones = phones;
-        this.ignore = ignore;
-    }
-
-    @Override
-    public String toString() {
-        return  "Person{" + "id=" + id +
-                ", name='" + name + '\'' +
-                ", age=" + age +
-                ", phones=" + phones +
-                ", ignore='" + ignore + '\'' +
-                '}';
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (!(o instanceof Person)) {
-            return false;
-        }
-        Person person = (Person) o;
-        return Objects.equals(id, person.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(id);
-    }
-
-    public static PersonBuilder builder() {
-        return new PersonBuilder();
-    }
-}
