--- artemis/artemis-graph/src/test/java/org/eclipse/jnosql/artemis/graph/model/User.java
@@ -1,91 +1,0 @@
-/*
- *  Copyright (c) 2017 Otávio Santana and others
- *   All rights reserved. This program and the accompanying materials
- *   are made available under the terms of the Eclipse Public License v1.0
- *   and Apache License v2.0 which accompanies this distribution.
- *   The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v10.html
- *   and the Apache License v2.0 is available at http://www.opensource.org/licenses/apache2.0.php.
- *
- *   You may elect to redistribute this code under either of these licenses.
- *
- *   Contributors:
- *
- *   Otavio Santana
- */
-package org.eclipse.jnosql.artemis.graph.model;
-
-import jakarta.nosql.mapping.Id;
-import org.apache.commons.lang3.builder.ToStringBuilder;
-
-import java.util.Objects;
-
-import static org.apache.commons.lang3.builder.ToStringStyle.SHORT_PREFIX_STYLE;
-
-
-public class User {
-
-    @Id
-    private String nickname;
-
-    private String name;
-
-    private int age;
-
-    public User(String nickname, String name, int age) {
-        this.nickname = nickname;
-        this.name = name;
-        this.age = age;
-    }
-
-    User() {}
-
-    public String getNickname() {
-        return nickname;
-    }
-
-    public void setNickname(String nickname) {
-        this.nickname = nickname;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        User user = (User) o;
-        return Objects.equals(nickname, user.nickname);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(nickname);
-    }
-
-    @Override
-    public String toString() {
-        return new ToStringBuilder(this, SHORT_PREFIX_STYLE)
-                .append("nickname", nickname)
-                .append("name", name)
-                .append("age", age)
-                .toString();
-    }
-}
