--- artemis/artemis-graph/src/test/java/org/eclipse/jnosql/artemis/graph/model/Vendor.java
@@ -1,62 +1,0 @@
-/*
- *
- *  Copyright (c) 2019 Otávio Santana and others
- *   All rights reserved. This program and the accompanying materials
- *   are made available under the terms of the Eclipse Public License v1.0
- *   and Apache License v2.0 which accompanies this distribution.
- *   The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v10.html
- *   and the Apache License v2.0 is available at http://www.opensource.org/licenses/apache2.0.php.
- *
- *   You may elect to redistribute this code under either of these licenses.
- *
- *   Contributors:
- *
- *   Otavio Santana
- *
- */
-package org.eclipse.jnosql.artemis.graph.model;
-
-import jakarta.nosql.mapping.Column;
-import jakarta.nosql.mapping.Entity;
-import jakarta.nosql.mapping.Id;
-
-import java.util.Collections;
-import java.util.Set;
-
-@Entity("vendors")
-public class Vendor {
-
-    @Id
-    private String name;
-
-    @Column
-    private Set<String> prefixes;
-
-    Vendor() {
-    }
-
-    public Vendor(String name) {
-        this.name = name;
-        prefixes = Collections.emptySet();
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Set<String> getPrefixes() {
-        return prefixes;
-    }
-
-    public void setPrefixes(Set<String> prefixes) {
-        this.prefixes = prefixes;
-    }
-
-    public void add(String prefix) {
-        this.prefixes.add(prefix);
-    }
-}
