--- artemis/artemis-graph/src/test/java/org/eclipse/jnosql/artemis/graph/model/Worker.java
@@ -1,60 +1,0 @@
-/*
- *  Copyright (c) 2017 Otávio Santana and others
- *   All rights reserved. This program and the accompanying materials
- *   are made available under the terms of the Eclipse Public License v1.0
- *   and Apache License v2.0 which accompanies this distribution.
- *   The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v10.html
- *   and the Apache License v2.0 is available at http://www.opensource.org/licenses/apache2.0.php.
- *
- *   You may elect to redistribute this code under either of these licenses.
- *
- *   Contributors:
- *
- *   Otavio Santana
- */
-package org.eclipse.jnosql.artemis.graph.model;
-
-
-import jakarta.nosql.mapping.Column;
-import jakarta.nosql.mapping.Convert;
-import jakarta.nosql.mapping.Entity;
-
-@Entity
-public class Worker {
-
-    @Column
-    private String name;
-
-    @Column
-    private Job job;
-
-    @Column("money")
-    @Convert(MoneyConverter.class)
-    private Money salary;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Job getJob() {
-        return job;
-    }
-
-    public void setJob(Job job) {
-        this.job = job;
-    }
-
-    public Money getSalary() {
-        return salary;
-    }
-
-    public void setSalary(Money salary) {
-        this.salary = salary;
-    }
-
-
-}
