+++ artemis/artemis-graph/src/test/java/org/eclipse/jnosql/artemis/graph/query/GraphRepositoryProxyPaginationTest.java
@@ -18,13 +18,13 @@ import jakarta.nosql.mapping.Converters;
 import jakarta.nosql.mapping.Pagination;
 import jakarta.nosql.mapping.Repository;
 import jakarta.nosql.mapping.reflection.ClassMappings;
+import jakarta.nosql.tck.entities.Person;
 import org.apache.tinkerpop.gremlin.structure.Edge;
 import org.apache.tinkerpop.gremlin.structure.Graph;
 import org.apache.tinkerpop.gremlin.structure.T;
 import org.apache.tinkerpop.gremlin.structure.Vertex;
 import org.eclipse.jnosql.artemis.graph.GraphConverter;
 import org.eclipse.jnosql.artemis.graph.GraphTemplate;
-import org.eclipse.jnosql.artemis.graph.model.Person;
 import jakarta.nosql.tck.test.CDIExtension;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -75,8 +75,6 @@ public class GraphRepositoryProxyPaginationTest {
 
         GraphRepositoryProxy personHandler = new GraphRepositoryProxy(template,
                 classMappings, PersonRepository.class,graph, converter, converters);
-
-
 
         when(template.insert(any(Person.class))).thenReturn(Person.builder().build());
         when(template.update(any(Person.class))).thenReturn(Person.builder().build());
